Fixed issue 12(1)
- Fixed issue 12(1)
New animated GIF behaviors #12
(https://github.com/TheLoombot/99-reddits/issues/12)

1. Sometimes animated GIFs are referenced but the URL filename will end
in ".jpg" ... here is an example: http://i.imgur.com/Qgnuarm.jpg ...
For these links, I found that it works if "gif" appears in the post
caption, but not otherwise. They should also work in the case where the
sub-reddit name itself has "gif" in it... Check out
/r/TestingPlatformGIF for this example.
2. Indirect references to the GIF files also have the same behavior:
They work if "gif" is in the caption title, but not in the sub-reddit
name. Check out /r/TestingPlatformGIF for this example.