Fixed issue #16
- Fixed issue #16:
Improve sharing of animated GIFs #16
(https://github.com/TheLoombot/99-reddits/issues/16)

One tiny change: Now that I see it, I realize that for messaging the
caption and URL are better left out. Please remove the post caption and
URL strings from the content when using the "Message" activity. It
should just be the image, no text. (It already behaves this way on
Copy.)